Unload global values (global variables)
When executable code is generated from the llvm::Module the JIT stores the
GlobalValue, Address in a map. This map is used to 'lock' the GlobalValue, so
that it doesn't go out of sync. If we want to 'unload' it we first need to
update the global mapping so that the JIT/ExecutionEngine 'forgets' about the
mapping.