The constructors need to be handled differently.
The constructors can be many kinds. For example, complete, base, vbase. Those
kinds are mangled in the symbol name. However the information about the constructor
kind is contained not in the ctor decl, but in the CXXConstructorExpr, attached
to a VarDecl, for example. This makes the information hard to retrieve. Instead
we just brute-force the module by removing all ctors of all possible kinds.

This in general is safe, because when we end up removing a ctor it comes from
removing the whole class.