Add TViewPubFunctions (view of all public functions).
This class is used in the implementation of TClass::GetListOfAllPublicMethods
and rather than copying the partial content of the ListOfMethods it presents
a view over theses list of the class and each of its bases classes.  By
definition any updated to the list of methods is immediately reflected in the
view.

The iteration go first through the public method of the class
then through the public method of its base classe (and the base
class's base classes) starting from the left most base class.

The view support iteration via TIter and direct access [ At(idx),
FindObject(TObject*) and FindObject(const char*) ].
The view does *not* supprot iteration through TObjLink and does *not* support
addition or removal of any content.