ROOT-5657 Fixed with "pragma clang ignored"s
Temporary fix, waiting for the pcms.
Recover old genreflex behaviour, i.e. do not print warnings due to glitches
in the headers at runtime. This is not synonym of ignoring warnings as they
will be printed at dictionary generation time.
In order to do this we leverage the diagnostic pragmas and, since there is no
way to express as a pragma the option "-Wno-deprecated" the
_BACKWARD_BACKWARD_WARNING_H macro, used to avoid to go through
backward/backward_warning.h.
The definition of the preprocessor macro ROOT_DICT_PRINT_WARNINGS re-enables
the warning printing at runtime. It is supposed to be used only for debugging
purposes.