Add TInterpreter::GetFunctionTemplate and FuncTempInfo_t support.
FuncTempInfo_t represent a Function Template Declaration.
We preserve the pattern of the other Info (i.e. it looks like to
the user as an opaque class that can be constructed and deleted)
but the implementation is simply to return (and never allocated
or dealloacte directly) the clang decl address.  Thus all
the implmenentation of 'TClingFunctionTemplate' is currently
directly into the implementation of the TCling::FuncTempInfo_*.