Add TBuffer::ReadVersionNoCheckSum
This (new) routine can be used from custom streamers
when it is known that the class was always versioned
(and thus never stored a checksum within the buffer).
This allows to disambiguate the case where the class
used to have a version number equal to zero and did
not save a byte count and is now versioned.

When reading the version number being zero, without
the byte count we have no way to guess whether the
class was version un-versioned or had previously a
version number equal to zero.