RooKeysPdf: faster initialisation of the lookup table
The construction of a RooKeysPdf is now much faster for large data sets.
This is achieved by carefully analysing where the different Gaussians
that the code superimposes can actually contribute. If a contribution
is smaller than the machine precision, the code stops trying to put in
those contributions (it would be lost in rounding anyway), which leads
to roughly 40 % faster construction of the class for large data sets.