Move function-static init check into unnamed namespace; spelling.
Multiple dictionaries with identical names caused the check to fail (only the first was ever initialized), as the symbol name was identical for all translation units.
Moving this into an unnamed namespace makes the symbol unique.
The "publicly" visible function (as used by TGClient) obviously still requires to be unique.