Implement a proper llvm::sys::path::reverse_iterator.
Since llvm::sys::path::const_iterator::operator* returns a reference to
a member of the iterator and std::reverse_iterator<T>::operator* call
T::operator* on a temporary object, llvm::sys::path:const_iterator can
not be used with std::reverse_iterator and we need to implement the reverse
iterator explicitly.

This fixes the execution of ROOT::TMetaUtils::GetFileName.