Improvement of dictionary generation
Now GenerateInitInstance is not created in presence of an std class
which is not a string.
This is achieved checking if the decl for which the dictionary is
created is of a certain type, expressed by a string.
This is more robust than any string manipulation.
As for the performance, a more aggressive implementation could
be put in place with for example a hashed map (unordered_map in
c++11) with as key the names of types and as vals the pointer to
the CXXRecordDecls corresponding to the actual definition of the
class.