Fix ROOT-5756 treating also names of typedefs/aliases of classes
This is obtained with the implementation of the VisitTypedefNameDecl
of the RScanner class, which is a RecursiveASTVisitor.
Basically the typedefs / c++11 aliases (e.g. using myAlias = int;)
are followed up, leveraging the TypdedefNameDecl class.
If they are selected with a selection xml, they refer to a
RecordDecl and they are not ultimately part of the namespace
std, they are selected.
The SelectionRules class had also to be massaged to accomodate the
treatment of the TypedefNameDecls.