Properly detect class layout match with array and counter variable.
In the case of array data member (both variable size and fixed size),
CompareContent was incorrectly reporting mismatches.

In the case of a counter variable, whether it is marked as a counter
variable or a regular variable might depends on the derived class used.
(This is the case for TCollection where fSize is used as to hold the
size of a variable size array only in some cases).  This results in
different files containing a different TStreamerInfo (for TCollection)
depending on what collections were stored in the file.  CompareContent
was upgraded to correctly ignore the difference.

This fixes ROOT-5151.