Add a class autoload callback.
Provide a way for a user of the ROOT library to receive a
callback during class autoloading.  TCling got a new data
member fAutoLoadCallback, and a way to get and set it.

The AutoLoad function will call the callback to give the
user a chance to load the library for the class itself.
If it does not, then the normal autoloading activities
will occur.