Use rootcling to generate rootmaps during the root build
To achieve this goal the following changes were applied:
1) rlibmap is reduced to a no-op with a nice deprecation message.
The return value is zero not to break build systems which may
rely on it. This will be transformed to an error in beta 2.
rlibmap will be removed in beta 3.
2) The build systems were updated in order to leverage rootcling
for the rootmap generation, basically adding 2 flags, -rmf and -rml.
3) Rootcling has a new option, -r which is on the same footing of
-f but it tells the system to ignore the dictionary generation.
When -r is present rootcling becomes a tool to generate rootmaps
(and capability files).

At the moment the makefile build does not properly account for
rootmaps as targets, i.e. a deleted/modified rootmap will trigger
a no-op rlibmap. A forthcoming commit will fix this behaviour.