Add code for backward comp. for broken case of abstract classes.
The code would be necessary only in very rare case and is added
here (but hidden behind an ifdef) in case this becomes a blocker.
One of the case would an abstract class that was written to
ROOT file older than v5.18 and in a circurcunstance that led
to an empty TStreamerInfo to be stored in the file and
if in addition one of the data member was declared using a
typedef _and_ the current class definition uses a different typedef.