Consistenly pass transient information through ShowMember
Update the ShowMembers and InspectMember set of functions to improve const
correctness and to pass through the information on whether the data analyzed
is transient or not (this allow suppression of diagnosis and less searches
in the transient case).

This is necessary to solve 'new' problems that appeared due to the fact that
we now have all the type informaiton rather than being limited to the explicit
selection of the user or by Cintex which was not passing on the transient
member information in the general case.  In addition genreflex does not
automatically generate the dictionary for STL collections that are used
(so ShowMember now sees thoses classes and BuildRealData was attempt
to create those objects eventhough they might not be creatable).

An example is the case of a map<A,B> where either A or B does not have default
constructor and thus the compilation of the default constructor for pair<A,B>
will fail (noisily) [This could also apply to any template instance, where it
might have a default constructor definition that can not be compiled due to
the template parameter]

This fixes ROOT-5880.