Made TClass::fgCallingNew a thread_local file scope static
In order to avoid thread-safety issues, the static class member
TClass::fgCallingNew is no longer a class member and is instead
a file scoped static declared thread_local. It was necessary to
not have it is a class member since CINT could not parse the new
thread_local keyword.

Conflicts:
	core/meta/inc/TClass.h