Clarification of offset direction, object type (ROOT-5597 and ROOT-5840).
For veritual inhertiance, the type of the object was ambiguous; it can be
either derived (as commonly used for ROOT and thus default) or base.
The latter is used by PyROOT's automatic upcast. The offset functions now
allow to specify the type (derived / base) of the object that is passed in.

"Derived" and "base" do not clarify the direction of the offset; they have
been replaced by "fromDerived" and "toBase" to signal the offset's sign.