Adapt to new TClass::GetClass(typeinfo)
Replace/Split the test for a completely missing dictionary for a data member
(which is now impossible) with a test for a sub-object of a class with a
non-requested dictionary (i.e. known to the interpreter but with no TClass
init) and a test for a object of a derived class only refered to by a pointer
to the base class (and which is really not known to the interpreter).