Add the notion of State to TClass.
This let us be more precise (and simplify the code asking the question)
about the current state of the TClass object.

   enum EState {
      kNoInfo,          // The state has not yet been initialized, i.e. the TClass
                        // was just created and/or there is no trace of it in the interpreter.
      kForwardDeclared, // The interpreted knows the entity is a class but that's it.
      kEmulated,        // The information about the class only comes from a TStreamerInfo
      kInterpreted,     // The class is described completely/only in the interpreter database.
      kHasTClassInit,   // The class has a TClass proper bootstrap coming from a run
                        // through rootcling/genreflex/TMetaUtils and the library
                        // containing this dictionary has been loaded in memory.
      kLoaded = kHasTClassInit
   };

This introduces in particular the state handled by kForwardDeclared,
while the state kEmulated/kInterpreted/kHasTClassInit could be calculated
from combination of IsLoaded and GetClassInfo.

This deprecates the use of the bit kIsEmulation.

Introduce a new constructor TClass(const char *name, Version_t cversion, Bool_t silent = kFALSE);
to be used from creating TClass object when not having much information.

The transition to and from kNoInfo,kForwardDeclared and kEmulated is now handled by
TClass::RegisterStreamerInfo and TClass::RemoveStreamerInfo.

Transition to kInterpreted is handled by TInterpreter::SetClassInfo.

Transition to kHasTClassInit is handled by the full fledge constructor.

We rely on the state kHasTClassInit to avoid generating the emulated collection proxy
when we know we will get compiled one a little bit later.