In TClassTable::Add register class/dict that must be run.
Whenever TClassTable::Add notices that there is already a TClass and
that this class has a ClassInfo (i.e. it will not be a new decl as
far as the interpreter is concerned), we register the TClass and
its dictionary function to be processed at the end of RegisterModule.

Add TInterpreter::RegisterTClassUpdate.

Note: This mechanism can not be a full replacement for the UpdateClassInfo
mechanism as some of the new decl loaded might not have a TClass initialization
(user include the header file but did not request the TClass initialization).