'Quick' fix for ROOT-5876 (and related issues)
If you feels like this is using a bandages to fix a leaking dam, you are likely correctly.

More technically, this avoid the routine that use GetInterpreterTypeName from re-expanding
the name (during the overly complex and obsolete scan through the list of possible names).

This avoids infinite recurse for vector<Long64_t> and map<Type,ABC_t>