Search the symbol in all available dylibs (ROOT-6015).
During static init the JIT could not find symbols in a library that the currently loaded library depends on.
This was caused by cling not having updated the list of loaded libraries yet - and only that list was searched for symbols.
Instead, now the dl handle 0 is searched, using the supposedly optimized symbol lookup tables of the process.
This leaves symbol priorities / lookup context to be dealt with: a library might not see the same symbols as the process.