Delegate the Update of the TClass to the CollectionProxy.
This allows to solved ROOT-6040 by avoid to have to initialize
the collection proxy just of find that ... we don't need to update it.

Having the useless update can lead to nested dlopen.  When a dlopen
loads a dictionary for a TClass that was already created (in emulated mode),
it will tell all the other TClass and all the CollectionProxy to update
the possible reference.  If at that point we initialize a collection proxy
which content is a represented by a TClass that has not yet been loaded
but has a rootmap file, we will end up with a nested call to dlopen to loaded
the library.