Simplify the update of the lookup tables when visiting a redeclarable.
The key idea is that the decl that is being detached from the redecl chain can
be in registered in the lookup tables. In this case we need to check if there
is another decl in the chain and update the lookup table so that the existing
decls in the chain could be accessible.