If the lookup vector doesn't contain decls remove the entry from the map.
The lookup table is highly optimized: a single data member can be a single decl
or a vector with decls. In the regular case when clang only adds declarations
the single decl gets promoted to a vector.

The decl reverter doesn't bother to demote the vector into a single decl (for now)
Insted if we see the vector is empty we remove the lookup entry from the map and
thus we make the compareState happy.