Fix ROOT-6036 (thanks bill)
Now IO read rules' c++ code fragments are compared differently.
The identity is not required anymore, rather the identity modulo
3 characters, '\n', ' ' and '\t'. Indeed their presence does not
alter the effect of the c++ code.
A possibility to improve the fix could be to generate a hash
starting from all the element of the io rules (modulo the
aforementioned manipulation of the code fragment!) and compare
just that one.