fgIsA needs to be thread safe
Classes which inherit from TObject have a fgIsA class member
defined by the code generator. However, the value for fgIsA is
set lazily and therefor must be protected in a threaded environment.
fgIsA was changed to be std::atomic<> and the setting of the value
was protected by gCINTMutex to avoid having the calculation done
more than once.

Conflicts:
	cint/reflex/python/genreflex/gendict.py
	core/base/inc/Rtypes.h
	core/utils/src/rootcint.cxx
	core/utils/src/rootcling.cxx