ROOT-5937: support ioname and iotype in selection xmls
The idea is to hardwire the iotype/ioname into the system.
This happens in 2 steps:
1) In the dictionaries, in code as we do for the "comment" field,
we can add attributes to the members for which an iotype/name is
specified in the selection xml. This will allow us to carry this
information around
2) In the TCling methods:
 - const char* TCling::DataMemberInfo_TypeTrueName(DataMemberInfo_t* dminfo) const
 - const char* TCling::DataMemberInfo_Name(DataMemberInfo_t* dminfo) const
 - const char* TCling::DataMemberInfo_TypeName(DataMemberInfo_t* dminfo) const
we can check which attributes we have and return the iotype/name if
we need.

The implementation is nothing to be proud about as it uses an internal cache.
It is not possible to use the buffer where the actual attribute is because
it may not be null terminated. So using llvm::stringRef::data() is not an option.
A review of the interfaces is of course possible as they are for internal use
only.