Add and use TClass::IsVersioned
'Foreign' has been meaning (for a while) only that the class does
not have a Streamer function and thus does not have a ClassDef.

This does not mean (and hasn't for a while) that the class layout
is not versioned as the version can be set externally via either
the linkdef or the selection or the RootClassVersion macro.

So add TClass::IsVersioned to clarify.

This fixes ROOT-6114

Conflicts:
	io/io/src/TStreamerInfo.cxx