Worked around ambiguity in C++ for fgIsA and std::atomic<>
The code created by genreflex was including the default namespace
qualifyer '::' on all class names. However, this lead to a C++
language ambiguity when parsing
   std::atomic<TClass*> ::Foo::fgIsA = 0
where the parser treats that as being
   std::atomic<TClass*>::Foo::fgIsA = 0
and fails since Foo is not a type defined in std::atomic<>.
The fix was to remove the '::' when generating the code. This then
matches what Cint does when generating the code.