Option for stressProof to print all logfiles
The new -catlog option prints all stressProof logfiles (including the
ones from the nodes/workers) on the stdout of the terminal used to
launch the test. This is useful i.e. for automatic tests (like CDash)
to have a look on what went wrong without logging on the test machine,
but just using the web interface. If -catlog is used with -cleanlog,
logs are displayed only if an error occurs.

The tests run with CDash have been updated to use this option.