Implement the iotype and ioname attributes for data members.
The implementations relies on these points:
 o The fieldDecl is decorated with an attribute carrying the information
   about iotype and ioname.
 o The TClingDataMemberInfo overrides the TypeName and Name with the
   values of the iotype and ioname if specified.

In addition, rootcling refuses (with a nice message) to assign an iotype
in case genreflex is used and the actual data member to be changed is a pod.

Add an optimised way to access properties' names and values has been added
to make the operations in TClingDataMemberInfo as fast as possible.