Prevent the possility of getting a TDataMember from a renamed member
when using the ioname property in the xml selection files.
The protection is placed in TCling::GetDataMember, where the call stack
ends up when requesting a datamember from TClass.
We check if the decl of the data member has an annotation which indicates
the presence of an ioname.
In case this is true, if the name requested is not the ioname, we
return 0, as if the member did not exist. In some sense we override
the information in the TClassInfo instance, isolating the typesystem in
TClass from the one in the AST.
In ROOT5 by Cintex was behind similar operations, i.e. building
a TClass entry from the information in Reflex.