Drops the validation of the directory name in prefetch
From Elvin:

I attached a patch which drops the validation of the directory name used for caching the prefetched blocks for the TFilePrefetching. I initially used it so that I wouldn't end up with funny names for the cache directory but the Alice guys (cc - Andrei) have some problems with this check when deploying it in production since they have various characters in the directory name which I currently don't accept.

Therefore, I completely removed this validation step and I rely on the TSystem/OS to report any errors if creating/opening the requested directory is not possible.