Reach the template decl within namespaces if present
the dictionary provides to RegisterModule a series of pairs of
string-uint. A pair is the template fwd declaration and the number
of template arguments to keep.
The system needs to discover in this fwd decl the template, in order
to keep a table of template ptr - # of templ args to keep.
For this reason, scopes are important.