Remove unused code.
TClassEdit is a string based cleaning of the names
and deal (mostly/mainly) with name of classes not known
to the interpreter.  As such it can not know which
(not-defined) namespace is inline or not and should
be removed (even if it is named "__1").  So we do
not (should not) hard code the name of any inline namespace.
Instead TClassEdit should always be given names where
the inline namespace have already been removed (because
they are either normalized by TInterpreter or comes
'somewhat normalized' from a file and a user written string.