Add checksum value to TStreamerBase.
In addition to the version number of the base class, record its checksum
to be able to properly handle the unversioned case.

Factorize the search for the StreamerInfo corresponding to the TStreamerBase
into it's initialization routine (In particular the
new TStreamerBase::InitStreaming).

At the same time add to this search support for renaming of the base class.

To store (in memory and on file) the value of the checksum we (ab)use the fact
that TStreamerElement always store 5 Int for the size of the array of the
element ... which is never used in the case of TStreamerBase.